requirements/dev.txt
-r base.txt
django-debug-toolbar==1.2.2
coverage==3.7.1
Fabric==1.8.2
isort==3.9.6
ipython==2.2.0